+++ src/main/kotlin/edu/unh/cs/ai/realtimesearch/planner/RealTimePlanner.kt
@@ -155,7 +155,7 @@ inline fun <StateType : State<StateType>, NodeType : RealTimeSearchNode<StateTyp
         }
 
         // Add the current state as the predecessor of the child state
-        successorNode.predecessors.add(SearchEdge(node = sourceNode, action = successor.action, actionCost = successor.actionCost))
+        successorNode.predecessors.add(SearchEdge(node = sourceNode, action = successor.action, actionCost = successor.actionCost.toLong()))
 
         // Skip if we got back to the parent
         if (successorState == sourceNode.parent.state) {
@@ -167,10 +167,10 @@ inline fun <StateType : State<StateType>, NodeType : RealTimeSearchNode<StateTyp
         if (successorNode.cost > successorGValueFromCurrent) {
             // here we generate a state. We store it's g value and remember how to get here via the treePointers
             successorNode.apply {
-                cost = successorGValueFromCurrent
+                cost = successorGValueFromCurrent.toLong()
                 parent = sourceNode
                 action = successor.action
-                actionCost = successor.actionCost
+                actionCost = successor.actionCost.toLong()
             }
 
             if (!successorNode.open) {
