+++ src/main/kotlin/edu/unh/cs/ai/realtimesearch/planner/realtime/SZero.kt
@@ -243,7 +243,7 @@ class SZeroPlanner<StateType : State<StateType>>(val domain: Domain<StateType>, 
             }
 
             // Add the current state as the predecessor of the child state
-            successorNode.predecessors.add(SearchEdge(node = sourceNode, action = successor.action, actionCost = successor.actionCost))
+            successorNode.predecessors.add(SearchEdge(node = sourceNode, action = successor.action, actionCost = successor.actionCost.toLong()))
 
             // Skip if we got back to the parent
             if (successorState == sourceNode.parent.state) {
@@ -255,10 +255,10 @@ class SZeroPlanner<StateType : State<StateType>>(val domain: Domain<StateType>, 
             if (successorNode.cost > successorGValueFromCurrent) {
                 // here we generate a state. We store it's g value and remember how to get here via the treePointers
                 successorNode.apply {
-                    cost = successorGValueFromCurrent
+                    cost = successorGValueFromCurrent.toLong()
                     parent = sourceNode
                     action = successor.action
-                    actionCost = successor.actionCost
+                    actionCost = successor.actionCost.toLong()
                 }
 
                 logger.debug { "Expanding from $sourceNode --> $successorState :: open list size: ${openList.size}" }
@@ -294,7 +294,7 @@ class SZeroPlanner<StateType : State<StateType>>(val domain: Domain<StateType>, 
             val undiscoveredNode = Node(
                     state = successorState,
                     heuristic = domain.heuristic(successorState),
-                    actionCost = successor.actionCost,
+                    actionCost = successor.actionCost.toLong(),
                     action = successor.action,
                     parent = parent,
                     cost = MAX_VALUE,
